requirements.txt
adjustText==1.2.0
anndata==0.8.0
bioinfokit==2.1.3
certifi==2023.7.22
cffi==1.16.0
click==8.1.3
cloudpickle==2.2.1
contourpy==1.0.6
cycler==0.11.0
Cython==3.0.3
dask==2023.6.1
dcor==0.6
distributed==2023.6.1
dtaidistance==2.3.10
fdasrsf==2.4.3
findiff==0.10.0
fonttools==4.38.0
fsspec==2023.6.0
future==0.18.2
h5py==3.7.0
igraph @ file:///home/conda/feedstock_root/build_artifacts/python-igraph_1696514381581/work
importlib-metadata==6.7.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.3
leidenalg @ file:///home/conda/feedstock_root/build_artifacts/leidenalg_1689886691453/work
llvmlite==0.39.1
locket==1.0.0
loess==2.1.2
MarkupSafe==2.1.3
matplotlib==3.6.2
matplotlib-venn==0.11.9
mpmath==1.3.0
msgpack==1.0.5
multimethod==1.10
natsort==8.2.0
networkx==2.8.8
numba==0.56.3
numpy==1.21.5
OneCC-bool==0.0.1
onesc==0.1.1
packaging==21.3
pandas==1.4.3
partd==1.4.0
patsy==0.5.3
Pillow==9.3.0
plotbin==3.1.5
progressbar2==4.2.0
psutil==5.9.5
pycparser==2.21
pyEpoch==0.1
pygad==3.1.0
pygam==0.8.0
pyitlib==0.2.3
pynndescent==0.5.7
pyparsing==3.0.9
python-dateutil==2.8.2
python-igraph==0.10.8
python-utils==3.4.5
pytz==2022.5
PyYAML==6.0
rdata==0.9
researchpy==0.3.5
scanpy==1.9.3
scikit-datasets==0.2.4
scikit-fda==0.8.1
scikit-learn==1.0.2
scipy==1.9.3
seaborn==0.11.2
session-info==1.0.0
six==1.16.0
sklearn==0.0.post9
sortedcontainers==2.4.0
statsmodels==0.13.2
stdlib-list==0.8.0
sympy==1.12
tabulate==0.9.0
tblib==2.0.0
texttable @ file:///home/conda/feedstock_root/build_artifacts/texttable_1696407540773/work
textwrap3==0.9.2
threadpoolctl==3.1.0
toolz==0.12.0
tornado==6.3.2
tqdm==4.64.1
typing_extensions==4.8.0
umap-learn==0.5.2
urllib3==2.0.3
xarray==2023.9.0
zict==3.0.0
zipp==3.15.0
